requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
psycopg2==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PySocks==1.7.1
pytest==8.0.2
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0